:construction: Extracts some behaviors
