+++ www/src/lib/lazyModules/tailwind.ts
@@ -1,0 +1,1 @@
+export const loadTailwind = () => import('cdn.tailwindcss.com/3.4.1?dlx');
